# refactor: Remove unnecessary codegen for UDFs and UDAFs.