--- a/ksql-engine/src/main/java/io/confluent/ksql/function/DynamicUdfInvoker.java
+++ b/ksql-engine/src/main/java/io/confluent/ksql/function/DynamicUdfInvoker.java
@@ -51,7 +51,8 @@ public class DynamicUdfInvoker implements UdfInvoker {
     try {
       final Object[] extractedArgs = extractArgs(args);
       for (int i = 0; i < extractedArgs.length; i++) {
-        extractedArgs[i] = UdfArgCoercer.coerceUdfArgs(args[i], method.getParameterTypes()[i], i);
+        extractedArgs[i] =
+            UdfArgCoercer.coerceUdfArgs(extractedArgs[i], method.getParameterTypes()[i], i);
       }
       return method.invoke(udf, extractedArgs);
     } catch (Exception e) {