### Description 

This refactoring considerably simplifies the UDF and UDAF creation code by removing codegen. 
Codegen was previously being used to wrap UDFs and UDAF factory methods in generated classes. This classes were then invoked dynamically to create the UDAF or invoke the UDF function.

The same behaviour can be achieved much more simply by just invoking the UDAF factory or UDF function directly, generating classes for this is unnecessary.

This PR mainly moves things around, renames a few things and deletes a significant amount of code. The PR shows a fair amount of new code but most of that is moved or cut and paste from other classes.

### Testing done 

Non functional change.

Updated and moved tests to reflect the new location and naming of the code.

### Reviewer checklist
- [ ] Ensure docs are updated if necessary. (eg. if a user visible feature is being added or changed).
- [ ] Ensure relevant issues are linked (description should include text like "Fixes #<issue number>")