@purplefox 2019-10-14
I've created a JMH benchmark and run it against master and this PR.
The benchmarks tests with a simple method with a single int param and a more complex method with varargs.
I also included benchmarks for just raw Method.invoke for those methods.
As expected performance is more or less same before and after. This should be no surprise as the execution is more or less the same (the old codegen just had a little more indirection), and both are using Method.invoke().
It's interesting that the timings for the invocation are much greater than raw Method.invoke - most of the time is actually taken on the arg munging and coercion that we do, not on the actual invoke(). (Actually there is scope to improve the coercion as we do a lot of that unnecessarily too - but that can be the subject of a different PR!).
Moral of the story - we don't really have to worry about the overhead of Method.invoke.

Master:
```
Benchmark                                Mode  Cnt  Score    Error  Units
UdfInvokerBenchmark.invokeSimple         avgt   12  7.041 ±  0.193  us/op
UdfInvokerBenchmark.invokeSimpleMethod   avgt   12  0.012 ±  0.001  us/op
UdfInvokerBenchmark.invokeVarargs        avgt   12  8.613 ±  0.362  us/op
```

This PR:
```
Benchmark                                Mode  Cnt  Score    Error  Units
UdfInvokerBenchmark.invokeSimple         avgt   12  7.397 ± 0.368  us/op
UdfInvokerBenchmark.invokeSimpleMethod   avgt   12  0.012 ± 0.001  us/op
UdfInvokerBenchmark.invokeVarargs        avgt   12  8.477 ± 0.521  us/op
UdfInvokerBenchmark.invokeVarargsMethod  avgt   12  0.013 ± 0.002  us/op
```