@purplefox 2019-10-14
Benchmark is here https://github.com/purplefox/ksql/pull/8/files