@big-andy-coates 2019-10-15
Awesome - thanks @purplefox.  I didn't realize that the old code used `Method.invoke()`, I was assumed avoid that was the reason for the code gen...